% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_oneToManyIntermodal.R
\name{mc_oneToManyIntermodal}
\alias{mc_oneToManyIntermodal}
\title{One to many routing
Computes the minimal duration from one place to many or vice versa.
The order in the response array corresponds to the order of coordinates of the \code{many} parameter in the query.}
\usage{
mc_oneToManyIntermodal(
  one = NULL,
  many = NULL,
  time = NULL,
  maxTravelTime = NULL,
  maxMatchingDistance = NULL,
  arriveBy = NULL,
  maxTransfers = NULL,
  minTransferTime = NULL,
  additionalTransferTime = NULL,
  transferTimeFactor = NULL,
  useRoutedTransfers = NULL,
  pedestrianProfile = NULL,
  pedestrianSpeed = NULL,
  cyclingSpeed = NULL,
  elevationCosts = NULL,
  transitModes = NULL,
  preTransitModes = NULL,
  postTransitModes = NULL,
  directMode = NULL,
  maxPreTransitTime = NULL,
  maxPostTransitTime = NULL,
  maxDirectTime = NULL,
  requireBikeTransport = NULL,
  requireCarTransport = NULL,
  .return_as = NULL,
  .json_parser = NULL,
  .headers = NULL,
  .auth = NULL,
  .throttle_rate = NULL,
  .build_only = NULL,
  .server = NULL,
  .endpoint = NULL,
  .referer = NULL,
  .req_options = NULL,
  .handle_response = NULL,
  .json_auto_unbox = NULL,
  .paginate = NULL
)
}
\arguments{
\item{one}{geo location as latitude;longitude}

\item{many}{geo locations as latitude;longitude,latitude;longitude,...

The number of accepted locations is limited by server config variable \code{onetomany_max_many}.}

\item{time}{Optional. Defaults to the current time.

Departure time ($arriveBy=false) / arrival date ($arriveBy=true),}

\item{maxTravelTime}{The maximum travel time in minutes.
If not provided, the routing uses the value
hardcoded in the server which is usually quite high.

\emph{Warning}: Use with care. Setting this too low can lead to
optimal (e.g. the least transfers) journeys not being found.
If this value is too low to reach the destination at all,
it can lead to slow routing performance.}

\item{maxMatchingDistance}{maximum matching distance in meters to match geo coordinates to the street network}

\item{arriveBy}{Optional. Defaults to false, i.e. one to many search

true = many to one
false = one to many}

\item{maxTransfers}{The maximum number of allowed transfers (i.e. interchanges between transit legs,
pre- and postTransit do not count as transfers).
\code{maxTransfers=0} searches for direct transit connections without any transfers.
If you want to search only for non-transit connections (\code{FOOT}, \code{CAR}, etc.),
send an empty \code{transitModes} parameter instead.

If not provided, the routing uses the server-side default value
which is hardcoded and very high to cover all use cases.

\emph{Warning}: Use with care. Setting this too low can lead to
optimal (e.g. the fastest) journeys not being found.
If this value is too low to reach the destination at all,
it can lead to slow routing performance.}

\item{minTransferTime}{Optional. Default is 0 minutes.

Minimum transfer time for each transfer in minutes.}

\item{additionalTransferTime}{Optional. Default is 0 minutes.

Additional transfer time reserved for each transfer in minutes.}

\item{transferTimeFactor}{Optional. Default is 1.0

Factor to multiply minimum required transfer times with.
Values smaller than 1.0 are not supported.}

\item{useRoutedTransfers}{Optional. Default is \code{false}.

Whether to use transfers routed on OpenStreetMap data.}

\item{pedestrianProfile}{Optional. Default is \code{FOOT}.

Accessibility profile to use for pedestrian routing in transfers
between transit connections and the first and last mile respectively.

Allowed values: FOOT, WHEELCHAIR.}

\item{pedestrianSpeed}{Optional

Average speed for pedestrian routing.}

\item{cyclingSpeed}{Optional

Average speed for bike routing.}

\item{elevationCosts}{Optional. Default is \code{NONE}.

Set an elevation cost profile, to penalize routes with incline.
\itemize{
\item \code{NONE}: No additional costs for elevations. This is the default behavior
\item \code{LOW}: Add a low cost for increase in elevation and incline along the way. This will prefer routes with less ascent, if small detours are required.
\item \code{HIGH}: Add a high cost for increase in elevation and incline along the way. This will prefer routes with less ascent, if larger detours are required.
}

As using an elevation costs profile will increase the travel duration,
routing through steep terrain may exceed the maximal allowed duration,
causing a location to appear unreachable.
Increasing the maximum travel time for these segments may resolve this issue.

The profile is used for routing on both the first and last mile.

Elevation cost profiles are currently used by following street modes:
\itemize{
\item \code{BIKE}
}

Allowed values: NONE, LOW, HIGH.}

\item{transitModes}{Optional. Default is \code{TRANSIT} which allows all transit modes (no restriction).
Allowed modes for the transit part. If empty, no transit connections will be computed.
For example, this can be used to allow only \verb{SUBURBAN,SUBWAY,TRAM}.

Allowed values: WALK, BIKE, RENTAL, CAR, CAR_PARKING, CAR_DROPOFF, ODM, RIDE_SHARING, FLEX, DEBUG_BUS_ROUTE, DEBUG_RAILWAY_ROUTE, DEBUG_FERRY_ROUTE, TRANSIT, TRAM, SUBWAY, FERRY, AIRPLANE, BUS, COACH, RAIL, HIGHSPEED_RAIL, LONG_DISTANCE, NIGHT_RAIL, REGIONAL_FAST_RAIL, REGIONAL_RAIL, SUBURBAN, FUNICULAR, AERIAL_LIFT, OTHER, AREAL_LIFT, METRO, CABLE_CAR.}

\item{preTransitModes}{Optional. Default is \code{WALK}. Does not apply to direct connections (see \code{directMode}).

A list of modes that are allowed to be used for the first mile, i.e. from the coordinates to the first transit stop. Example: \verb{WALK,BIKE_SHARING}.

Allowed values: WALK, BIKE, RENTAL, CAR, CAR_PARKING, CAR_DROPOFF, ODM, RIDE_SHARING, FLEX, DEBUG_BUS_ROUTE, DEBUG_RAILWAY_ROUTE, DEBUG_FERRY_ROUTE, TRANSIT, TRAM, SUBWAY, FERRY, AIRPLANE, BUS, COACH, RAIL, HIGHSPEED_RAIL, LONG_DISTANCE, NIGHT_RAIL, REGIONAL_FAST_RAIL, REGIONAL_RAIL, SUBURBAN, FUNICULAR, AERIAL_LIFT, OTHER, AREAL_LIFT, METRO, CABLE_CAR.}

\item{postTransitModes}{Optional. Default is \code{WALK}. Does not apply to direct connections (see \code{directMode}).

A list of modes that are allowed to be used for the last mile, i.e. from the last transit stop to the target coordinates. Example: \verb{WALK,BIKE_SHARING}.

Allowed values: WALK, BIKE, RENTAL, CAR, CAR_PARKING, CAR_DROPOFF, ODM, RIDE_SHARING, FLEX, DEBUG_BUS_ROUTE, DEBUG_RAILWAY_ROUTE, DEBUG_FERRY_ROUTE, TRANSIT, TRAM, SUBWAY, FERRY, AIRPLANE, BUS, COACH, RAIL, HIGHSPEED_RAIL, LONG_DISTANCE, NIGHT_RAIL, REGIONAL_FAST_RAIL, REGIONAL_RAIL, SUBURBAN, FUNICULAR, AERIAL_LIFT, OTHER, AREAL_LIFT, METRO, CABLE_CAR.}

\item{directMode}{Default is \code{WALK} which will compute walking routes as direct connections.

Mode used for direction connections from start to destination without using transit.

Currently supported non-transit modes: \code{WALK}, \code{BIKE}, \code{CAR}

Allowed values: WALK, BIKE, RENTAL, CAR, CAR_PARKING, CAR_DROPOFF, ODM, RIDE_SHARING, FLEX, DEBUG_BUS_ROUTE, DEBUG_RAILWAY_ROUTE, DEBUG_FERRY_ROUTE, TRANSIT, TRAM, SUBWAY, FERRY, AIRPLANE, BUS, COACH, RAIL, HIGHSPEED_RAIL, LONG_DISTANCE, NIGHT_RAIL, REGIONAL_FAST_RAIL, REGIONAL_RAIL, SUBURBAN, FUNICULAR, AERIAL_LIFT, OTHER, AREAL_LIFT, METRO, CABLE_CAR.}

\item{maxPreTransitTime}{Optional. Default is 15min which is \code{900}.
Maximum time in seconds for the first street leg.
Is limited by server config variable \code{street_routing_max_prepost_transit_seconds}.}

\item{maxPostTransitTime}{Optional. Default is 15min which is \code{900}.
Maximum time in seconds for the last street leg.
Is limited by server config variable \code{street_routing_max_prepost_transit_seconds}.}

\item{maxDirectTime}{Optional. Default is 30min which is \code{1800}.
Maximum time in seconds for direct connections.

If a value smaller than either \code{maxPreTransitTime} or
\code{maxPostTransitTime} is used, their maximum is set instead.
Is limited by server config variable \code{street_routing_max_direct_seconds}.}

\item{requireBikeTransport}{Optional. Default is \code{false}.

If set to \code{true}, all used transit trips are required to allow bike carriage.}

\item{requireCarTransport}{Optional. Default is \code{false}.

If set to \code{true}, all used transit trips are required to allow car carriage.}

\item{.return_as}{A string specifying the return format. Defaults to 'list'. Options are 'list' for a parsed R list, 'raw' for the raw httr2_response object, or 'string' for the raw JSON string.}

\item{.json_parser}{A string specifying which parser to use when .return_as = 'list'. Defaults to 'RcppSimdJson' (faster) or 'jsonlite'. Beware that their output may differ slightly.}

\item{.headers}{A named list of extra HTTP headers to add to the request. If no 'Accept' header is provided anywhere, 'Accept: \emph{/}' will be sent by default.}

\item{.auth}{An authentication token or method (e.g., a bearer token string).}

\item{.throttle_rate}{A number to pass to \code{httr2::req_throttle()} to limit the rate of requests. For example, \code{15/60} means 15 requests per minute. This overrides any default rate set at client generation time.}

\item{.build_only}{A logical. If TRUE, the function will build and return the httr2_request object without performing it. Defaults to FALSE.}

\item{.server}{A string to override the base URL for this specific request. If provided, it will be used instead of the default server URL.}

\item{.endpoint}{A string to override the path for this specific request. If provided, it will be used instead of the default path from the specification. For example, use '/v2/users' to override the endpoint path.}

\item{.referer}{A string to set the Referer HTTP header.}

\item{.req_options}{A named list of curl options passed to \code{httr2::req_options()}. Useful for timeouts, proxies, SSL, etc.}

\item{.handle_response}{A function taking an \code{httr2_response} and returning a value. If supplied, it overrides \code{.return_as} / \code{.json_parser} handling. See helpers like \code{oa3_content_or_stop()}.}

\item{.json_auto_unbox}{Logical. If TRUE, JSON bodies are encoded with \code{auto_unbox = TRUE} (jsonlite). Defaults to FALSE unless overridden.}

\item{.paginate}{A logical, character string, or function to enable pagination. If TRUE (or "link_header"), uses Link headers. Other options: "page_param", "cursor_param", or a custom function. See \code{oa3_paginate()}.}
}
\description{
One to many routing
Computes the minimal duration from one place to many or vice versa.
The order in the response array corresponds to the order of coordinates of the \code{many} parameter in the query.
}
