% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_oneToManyIntermodalPost.R
\name{mc_oneToManyIntermodalPost}
\alias{mc_oneToManyIntermodalPost}
\title{One to many routing
Computes the minimal duration from one place to many or vice versa.
The order in the response array corresponds to the order of coordinates of the \code{many} parameter in the query.}
\usage{
mc_oneToManyIntermodalPost(
  one = NULL,
  many = NULL,
  time = NULL,
  maxTravelTime = NULL,
  maxMatchingDistance = NULL,
  arriveBy = NULL,
  maxTransfers = NULL,
  minTransferTime = NULL,
  additionalTransferTime = NULL,
  transferTimeFactor = NULL,
  useRoutedTransfers = NULL,
  pedestrianProfile = NULL,
  pedestrianSpeed = NULL,
  cyclingSpeed = NULL,
  elevationCosts = NULL,
  transitModes = NULL,
  preTransitModes = NULL,
  postTransitModes = NULL,
  directMode = NULL,
  maxPreTransitTime = NULL,
  maxPostTransitTime = NULL,
  maxDirectTime = NULL,
  requireBikeTransport = NULL,
  requireCarTransport = NULL,
  .body = NULL,
  .return_as = NULL,
  .json_parser = NULL,
  .headers = NULL,
  .auth = NULL,
  .throttle_rate = NULL,
  .build_only = NULL,
  .server = NULL,
  .endpoint = NULL,
  .referer = NULL,
  .req_options = NULL,
  .handle_response = NULL,
  .json_auto_unbox = NULL,
  .paginate = NULL
)
}
\arguments{
\item{one}{\if{html}{\out{<string>}} \verb{latitude,longitude[,level]} tuple with - latitude and longitude in degrees - (optional) level: the OSM level (default: 0) OR stop id}

\item{many}{\if{html}{\out{<array>}} array of: \verb{latitude,longitude[,level]} tuple with - latitude and longitude in degrees - (optional) level: the OSM level (default: 0) OR stop id The number of accepted locations is limited by server config variable \code{onetomany_max_many}.}

\item{time}{\if{html}{\out{<string>}} Optional. Defaults to the current time. Departure time ($arriveBy=false) / arrival date ($arriveBy=true),}

\item{maxTravelTime}{\if{html}{\out{<integer>}} The maximum travel time in minutes. If not provided, the routing uses the value hardcoded in the server which is usually quite high. \emph{Warning}: Use with care. Setting this too low can lead to optimal (e.g. the least transfers) journeys not being found. If this value is too low to reach the destination at all, it can lead to slow routing performance.}

\item{maxMatchingDistance}{\if{html}{\out{<number>}} maximum matching distance in meters to match geo coordinates to the street network}

\item{arriveBy}{\if{html}{\out{<boolean>}} Optional. Defaults to false, i.e. one to many search true = many to one false = one to many}

\item{maxTransfers}{\if{html}{\out{<integer>}} The maximum number of allowed transfers (i.e. interchanges between transit legs, pre- and postTransit do not count as transfers). \code{maxTransfers=0} searches for direct transit connections without any transfers. If you want to search only for non-transit connections (\code{FOOT}, \code{CAR}, etc.), send an empty \code{transitModes} parameter instead. If not provided, the routing uses the server-side default value which is hardcoded and very high to cover all use cases. \emph{Warning}: Use with care. Setting this too low can lead to optimal (e.g. the fastest) journeys not being found. If this value is too low to reach the destination at all, it can lead to slow routing performance.}

\item{minTransferTime}{\if{html}{\out{<integer>}} Optional. Default is 0 minutes. Minimum transfer time for each transfer in minutes.}

\item{additionalTransferTime}{\if{html}{\out{<integer>}} Optional. Default is 0 minutes. Additional transfer time reserved for each transfer in minutes.}

\item{transferTimeFactor}{\if{html}{\out{<number>}} Optional. Default is 1.0 Factor to multiply minimum required transfer times with. Values smaller than 1.0 are not supported.}

\item{useRoutedTransfers}{\if{html}{\out{<boolean>}} Optional. Default is \code{false}. Whether to use transfers routed on OpenStreetMap data.}

\item{pedestrianProfile}{\if{html}{\out{<string>}} Different accessibility profiles for pedestrians.
Allowed values: FOOT, WHEELCHAIR.}

\item{pedestrianSpeed}{\if{html}{\out{<number>}} Average speed for pedestrian routing in meters per second}

\item{cyclingSpeed}{\if{html}{\out{<number>}} Average speed for bike routing in meters per second}

\item{elevationCosts}{\if{html}{\out{<string>}} Different elevation cost profiles for street routing. Using a elevation cost profile will prefer routes with a smaller incline and smaller difference in elevation, even if the routed way is longer. - \code{NONE}: Ignore elevation data for routing. This is the default behavior - \code{LOW}: Add a low penalty for inclines. This will favor longer paths, if the elevation increase and incline are smaller. - \code{HIGH}: Add a high penalty for inclines. This will favor even longer paths, if the elevation increase and incline are smaller.
Allowed values: NONE, LOW, HIGH.}

\item{transitModes}{\if{html}{\out{<array>}} Optional. Default is \code{TRANSIT} which allows all transit modes (no restriction). Allowed modes for the transit part. If empty, no transit connections will be computed. For example, this can be used to allow only \verb{SUBURBAN,SUBWAY,TRAM}.
Allowed values: WALK, BIKE, RENTAL, CAR, CAR_PARKING, CAR_DROPOFF, ODM, RIDE_SHARING, FLEX, DEBUG_BUS_ROUTE, DEBUG_RAILWAY_ROUTE, DEBUG_FERRY_ROUTE, TRANSIT, TRAM, SUBWAY, FERRY, AIRPLANE, BUS, COACH, RAIL, HIGHSPEED_RAIL, LONG_DISTANCE, NIGHT_RAIL, REGIONAL_FAST_RAIL, REGIONAL_RAIL, SUBURBAN, FUNICULAR, AERIAL_LIFT, OTHER, AREAL_LIFT, METRO, CABLE_CAR.}

\item{preTransitModes}{\if{html}{\out{<array>}} Optional. Default is \code{WALK}. Does not apply to direct connections (see \code{directMode}). A list of modes that are allowed to be used for the first mile, i.e. from the coordinates to the first transit stop. Example: \verb{WALK,BIKE_SHARING}.
Allowed values: WALK, BIKE, RENTAL, CAR, CAR_PARKING, CAR_DROPOFF, ODM, RIDE_SHARING, FLEX, DEBUG_BUS_ROUTE, DEBUG_RAILWAY_ROUTE, DEBUG_FERRY_ROUTE, TRANSIT, TRAM, SUBWAY, FERRY, AIRPLANE, BUS, COACH, RAIL, HIGHSPEED_RAIL, LONG_DISTANCE, NIGHT_RAIL, REGIONAL_FAST_RAIL, REGIONAL_RAIL, SUBURBAN, FUNICULAR, AERIAL_LIFT, OTHER, AREAL_LIFT, METRO, CABLE_CAR.}

\item{postTransitModes}{\if{html}{\out{<array>}} Optional. Default is \code{WALK}. Does not apply to direct connections (see \code{directMode}). A list of modes that are allowed to be used for the last mile, i.e. from the last transit stop to the target coordinates. Example: \verb{WALK,BIKE_SHARING}.
Allowed values: WALK, BIKE, RENTAL, CAR, CAR_PARKING, CAR_DROPOFF, ODM, RIDE_SHARING, FLEX, DEBUG_BUS_ROUTE, DEBUG_RAILWAY_ROUTE, DEBUG_FERRY_ROUTE, TRANSIT, TRAM, SUBWAY, FERRY, AIRPLANE, BUS, COACH, RAIL, HIGHSPEED_RAIL, LONG_DISTANCE, NIGHT_RAIL, REGIONAL_FAST_RAIL, REGIONAL_RAIL, SUBURBAN, FUNICULAR, AERIAL_LIFT, OTHER, AREAL_LIFT, METRO, CABLE_CAR.}

\item{directMode}{\if{html}{\out{<string>}} # Street modes - \code{WALK} - \code{BIKE} - \code{RENTAL} Experimental. Expect unannounced breaking changes (without version bumps) for all parameters and returned structs. - \code{CAR} - \code{CAR_PARKING} Experimental. Expect unannounced breaking changes (without version bumps) for all parameters and returned structs. - \code{CAR_DROPOFF} Experimental. Expect unannounced breaking changes (without version bumps) for all perameters and returned structs. - \code{ODM} on-demand taxis from the Prima+ÖV Project - \code{RIDE_SHARING} ride sharing from the Prima+ÖV Project - \code{FLEX} flexible transports # Transit modes - \code{TRANSIT}: translates to \verb{TRAM,FERRY,AIRPLANE,BUS,COACH,RAIL,ODM,FUNICULAR,AERIAL_LIFT,OTHER} - \code{TRAM}: trams - \code{SUBWAY}: subway trains (Paris Metro, London Underground, but also NYC Subway, Hamburger Hochbahn, and other non-underground services) - \code{FERRY}: ferries - \code{AIRPLANE}: airline flights - \code{BUS}: short distance buses (does not include \code{COACH}) - \code{COACH}: long distance buses (does not include \code{BUS}) - \code{RAIL}: translates to \verb{HIGHSPEED_RAIL,LONG_DISTANCE,NIGHT_RAIL,REGIONAL_RAIL,SUBURBAN,SUBWAY} - \code{HIGHSPEED_RAIL}: long distance high speed trains (e.g. TGV) - \code{LONG_DISTANCE}: long distance inter city trains - \code{NIGHT_RAIL}: long distance night trains - \code{REGIONAL_FAST_RAIL}: deprecated, \code{REGIONAL_RAIL} will be used - \code{REGIONAL_RAIL}: regional train - \code{SUBURBAN}: suburban trains (e.g. S-Bahn, RER, Elizabeth Line, ...) - \code{ODM}: demand responsive transport - \code{FUNICULAR}: Funicular. Any rail system designed for steep inclines. - \code{AERIAL_LIFT}: Aerial lift, suspended cable car (e.g., gondola lift, aerial tramway). Cable transport where cabins, cars, gondolas or open chairs are suspended by means of one or more cables. - \code{AREAL_LIFT}: deprecated - \code{METRO}: deprecated - \code{CABLE_CAR}: deprecated
Allowed values: WALK, BIKE, RENTAL, CAR, CAR_PARKING, CAR_DROPOFF, ODM, RIDE_SHARING, FLEX, DEBUG_BUS_ROUTE, DEBUG_RAILWAY_ROUTE, DEBUG_FERRY_ROUTE, TRANSIT, TRAM, SUBWAY, FERRY, AIRPLANE, BUS, COACH, RAIL, HIGHSPEED_RAIL, LONG_DISTANCE, NIGHT_RAIL, REGIONAL_FAST_RAIL, REGIONAL_RAIL, SUBURBAN, FUNICULAR, AERIAL_LIFT, OTHER, AREAL_LIFT, METRO, CABLE_CAR.}

\item{maxPreTransitTime}{\if{html}{\out{<integer>}} Optional. Default is 15min which is \code{900}. Maximum time in seconds for the first street leg. Is limited by server config variable \code{street_routing_max_prepost_transit_seconds}.}

\item{maxPostTransitTime}{\if{html}{\out{<integer>}} Optional. Default is 15min which is \code{900}. Maximum time in seconds for the last street leg. Is limited by server config variable \code{street_routing_max_prepost_transit_seconds}.}

\item{maxDirectTime}{\if{html}{\out{<integer>}} Optional. Default is 30min which is \code{1800}. Maximum time in seconds for direct connections. If a value smaller than either \code{maxPreTransitTime} or \code{maxPostTransitTime} is used, their maximum is set instead. Is limited by server config variable \code{street_routing_max_direct_seconds}.}

\item{requireBikeTransport}{\if{html}{\out{<boolean>}} Optional. Default is \code{false}. If set to \code{true}, all used transit trips are required to allow bike carriage.}

\item{requireCarTransport}{\if{html}{\out{<boolean>}} Optional. Default is \code{false}. If set to \code{true}, all used transit trips are required to allow car carriage.}

\item{.body}{A list or object to be sent as the request body. If NULL (the default), the body will be constructed from any other top-level arguments you provide.}

\item{.return_as}{A string specifying the return format. Defaults to 'list'. Options are 'list' for a parsed R list, 'raw' for the raw httr2_response object, or 'string' for the raw JSON string.}

\item{.json_parser}{A string specifying which parser to use when .return_as = 'list'. Defaults to 'RcppSimdJson' (faster) or 'jsonlite'. Beware that their output may differ slightly.}

\item{.headers}{A named list of extra HTTP headers to add to the request. If no 'Accept' header is provided anywhere, 'Accept: \emph{/}' will be sent by default.}

\item{.auth}{An authentication token or method (e.g., a bearer token string).}

\item{.throttle_rate}{A number to pass to \code{httr2::req_throttle()} to limit the rate of requests. For example, \code{15/60} means 15 requests per minute. This overrides any default rate set at client generation time.}

\item{.build_only}{A logical. If TRUE, the function will build and return the httr2_request object without performing it. Defaults to FALSE.}

\item{.server}{A string to override the base URL for this specific request. If provided, it will be used instead of the default server URL.}

\item{.endpoint}{A string to override the path for this specific request. If provided, it will be used instead of the default path from the specification. For example, use '/v2/users' to override the endpoint path.}

\item{.referer}{A string to set the Referer HTTP header.}

\item{.req_options}{A named list of curl options passed to \code{httr2::req_options()}. Useful for timeouts, proxies, SSL, etc.}

\item{.handle_response}{A function taking an \code{httr2_response} and returning a value. If supplied, it overrides \code{.return_as} / \code{.json_parser} handling. See helpers like \code{oa3_content_or_stop()}.}

\item{.json_auto_unbox}{Logical. If TRUE, JSON bodies are encoded with \code{auto_unbox = TRUE} (jsonlite). Defaults to FALSE unless overridden.}

\item{.paginate}{A logical, character string, or function to enable pagination. If TRUE (or "link_header"), uses Link headers. Other options: "page_param", "cursor_param", or a custom function. See \code{oa3_paginate()}.}
}
\description{
One to many routing
Computes the minimal duration from one place to many or vice versa.
The order in the response array corresponds to the order of coordinates of the \code{many} parameter in the query.
}
